Add files via upload
I have updated the "footer" element; to center-align the link to the "Top" of the page, changed the Glyphicon for Copyright to a Icon Font from "Font Awesome". The "fixed-bottom" does not swallow the buttons in the form now. 
I have also changed the Logo-picture in the Navbar at the top to shrink the menu somewhat.